+++ transport/handler_server.go
@@ -301,8 +301,8 @@ func (ht *serverHandlerTransport) HandleStreams(startStream func(*Stream), trace
 
 	s := &Stream{
 		id:            0,            // irrelevant
-		streamWindowHandler: func(uint32) {},
-		transportWindowHandler: func(uint32) {},
+		streamWindowHandler: func(int64) {},
+		transportWindowHandler: func(int64) {},
 		cancel:        cancel,
 		buf:           newRecvBuffer(),
 		st:            ht,
